refractoring response, download folder check.
